+++ pom.xml
@@ -96,7 +96,7 @@
                 </artifactSet>
                 <transformers>
                   <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                    <mainClass>com.xiaomi.infra.sample.spark.scanAllData</mainClass>
+                    <mainClass>com.xiaomi.infra.pegasus.analyser.sample.spark.scanAllData</mainClass>
                   </transformer>
                 </transformers>
               </configuration>
