+++ scripts/format-all.sh
@@ -4,11 +4,7 @@ SCRIPT_DIR=$(dirname "${BASH_SOURCE[0]}")
 PROJECT_DIR=$(dirname "${SCRIPT_DIR}")
 cd "${PROJECT_DIR}" || exit 1
 
-SRC_FILES=(src/main/java/com/xiaomi/infra/common/*.java
-           src/main/java/com/xiaomi/infra/config/*.java
-           src/main/java/com/xiaomi/infra/service/db/*.java
-           src/main/java/com/xiaomi/infra/service/*.java
-           src/main/java/com/xiaomi/infra/*.java
+SRC_FILES=(src/main/java/com/xiaomi/infra/pegasus/analyser/*.java
            )
 
 if [ ! -f "${PROJECT_DIR}"/google-java-format-1.7-all-deps.jar ]; then
