+++ src/main/java/com/xiaomi/infra/pegasus/analyser/PegasusScanner.java
@@ -1,6 +1,5 @@
-package com.xiaomi.infra.service.db;
+package com.xiaomi.infra.pegasus.analyser;
 
-import com.xiaomi.infra.common.Utils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.rocksdb.RocksDB;
 import org.rocksdb.RocksIterator;
