+++ src/main/java/com/xiaomi/infra/pegasus/analyser/sample/spark/countAllData.scala
@@ -1,12 +1,10 @@
-package com.xiaomi.infra.sample.spark
+package com.xiaomi.infra.pegasus.analyser.sample.spark
 
 import java.util.concurrent.atomic.AtomicInteger
 
-import com.xiaomi.infra.PegasusClient
-import com.xiaomi.infra.config.Config
-import com.xiaomi.infra.sample.local.ScanTool
-import com.xiaomi.infra.service.FdsService
-import com.xiaomi.infra.service.db.PegasusOptions
+import com.xiaomi.infra.pegasus.analyser.{Config, FdsService, PegasusClient, PegasusOptions}
+import com.xiaomi.infra.{FdsService, PegasusClient, PegasusOptions}
+import com.xiaomi.infra.pegasus.analyser.sample.local.ScanTool
 import org.apache.commons.logging.LogFactory
 import org.apache.spark.{SparkConf, SparkContext, TaskContext}
 import org.rocksdb.RocksDB
